[fix] risk: freeze-leak on ME send-fail, price/qty guard, index qty overflow, migration lock, log ANSI
- RISK-INDEX-QTY-OVERFLOW (price.rs): widen bid_qty/ask_qty to i128
  before summing (was i64 add -> debug panic / release wrap).
- RISK-FREEZE-LEAK-ON-ME-SEND-FAIL (main.rs): on ME send error or
  missing sender for an accepted order, release_frozen_for_order +
  reply ORDER_FAILED (NetworkError) to gateway; was log-only -> the
  in-memory freeze leaked margin forever.
- RISK-NO-PRICE-QTY-GUARD (shard.rs process_order): reject
  price<=0 || qty<=0 before margin math. Gateway already guards on
  the WS path (convert.rs price>0/qty>0); risk is the solvency
  authority -> defense in depth. Negative notional would yield a
  negative freeze that adds available margin.
- MIGRATIONS-UNLOCKED (main.rs): wrap run_migrations in
  pg_advisory_lock(MIGRATION_LOCK_KEY) (distinct negative sentinel
  key from per-shard shard_id) so exactly one node migrates under
  the eager-replica boot, where the main lock is won late.
- ANSI-IN-LOGS (main.rs): tracing fmt .with_ansi(is_terminal(stdout)).

Tests: +index_price_qty_sum_no_i64_overflow,
+order_nonpositive_price_or_qty_rejected. 288 pass / 0 fail.

+++ rsx-risk/src/price.rs
@@ -15,7 +15,7 @@ pub fn calculate_index(
     ask_qty: i64,
     last_index: i64,
 ) -> i64 {
-    let total = bid_qty + ask_qty;
+    let total = bid_qty as i128 + ask_qty as i128;
     if total == 0 {
         return last_index;
     }
@@ -27,7 +27,7 @@ pub fn calculate_index(
     }
     ((bid_px as i128 * ask_qty as i128
         + ask_px as i128 * bid_qty as i128)
-        / total as i128).clamp(i64::MIN as i128, i64::MAX as i128) as i64
+        / total).clamp(i64::MIN as i128, i64::MAX as i128) as i64
 }
 
 impl IndexPrice {
